fix: Fix unlink vendor account
