+++ src/user/user.controller.ts
@@ -1,10 +1,7 @@
-import { JwtAuthGuard } from '@/auth/jwt-auth.guard.js';
-import { Body, Controller, Delete, HttpCode, Param, Post, UseGuards } from '@nestjs/common';
-import { ApiBearerAuth, ApiCreatedResponse, ApiNoContentResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
-import { User } from '@prisma/client';
+import { Body, Controller, HttpCode, Post } from '@nestjs/common';
+import { ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { SignUpDto } from './dto/signup.dto.js';
 import { UserDto } from './dto/user.dto.js';
-import { UserAuthentication } from './user-authentication.decorator.js';
 import { UserService } from './user.service.js';
 
 @Controller('users')
@@ -19,15 +16,5 @@ export class UserController {
     async signup(@Body() signUpDto: SignUpDto) {
         const user = await this.userService.signup(signUpDto);
         return new UserDto(user.id, user.username);
-    }
-
-    @UseGuards(JwtAuthGuard)
-    @Delete('stream/account/:accountId')
-    @HttpCode(204)
-    @ApiBearerAuth('Authorization')
-    @ApiOperation({ summary: 'Unlink stream account API', description: 'Unlinks a stream account from the user' })
-    @ApiNoContentResponse()
-    async unlinkStreamAccount(@UserAuthentication() user: User, @Param('accountId') accountId: string) {
-        await this.userService.unlinkStreamAccount(user, accountId);
     }
 }
