+++ src/user/user.service.ts
@@ -1,6 +1,5 @@
-import { ConflictException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { User } from '@prisma/client';
-import dayjs from 'dayjs';
 import { VendorAccountRepository } from '../vendor-account/vendor-account.repository.js';
 import { SignUpDto } from './dto/signup.dto.js';
 import { UserRepository } from './user.repository.js';
@@ -19,21 +18,5 @@ export class UserService {
             throw new ConflictException('conflict', 'user already exists');
         }
         return this.userRepository.create({ id, username });
-    }
-
-    async unlinkStreamAccount(user: User, accountId: string): Promise<void> {
-        const vendorAccount = await this.vendorAccountRepository.findById(accountId).catch(() => {
-            throw new NotFoundException('Not Found', 'vendor account not found');
-        });
-        if (vendorAccount.userId !== user.id) {
-            throw new UnauthorizedException('Unauthorized', 'vendor account does not belong to user');
-        }
-
-        await this.vendorAccountRepository.update({
-            where: { id: vendorAccount.id },
-            data: { deactivatedAt: dayjs().toDate() },
-        });
-
-        return;
     }
 }
